refactor(dify-ui): consolidate form control styles
