--- packages/dify-ui/src/form-shared.ts
@@ -1,1 +1,0 @@
-export const formLabelClassName = 'w-fit py-1 text-text-secondary system-sm-medium data-disabled:cursor-not-allowed'
